requirements.txt
django 
djangorestframework
mysqlclient
